// src/__generated__/player.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: qUoSr1nTFiT5WHMbmcMCZHUE8myxVv9utUeepheMiDXXlKmbf6IBMdoxD2+BHSwBPDmmtVVeATQGsekpWs9v0w==
 */

/* eslint-disable */
// tslint:disable

interface Player {
  steam_id: bigint & {readonly __brand?: 'player_steam_id'}
}
export default Player;

interface Player_InsertParameters {
  steam_id: bigint & {readonly __brand?: 'player_steam_id'}
}
export type {Player_InsertParameters}